BACKEND/requirements.txt
flask
openai
sqlalchemy
pymysql
python-dotenv